chore(DEV-11865): more refactoring for architecture
